% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{areColours}
\alias{areColours}
\title{Colour vectors}
\usage{
areColours(x)
}
\arguments{
\item{x}{character vector of colours to test}
}
\value{
a boolean value
}
\description{
Test if an object is interpretable as colours.
}
