% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brewerPalInfo}
\alias{brewerPalInfo}
\title{Retrieve Information on RColorBrewer Palettes}
\usage{
brewerPalInfo()
}
\value{
dataframe with four columns: name, maxcolors, category, colorblind
}
\description{
Retrieve information on RColorBrewer palette lengths, categories and colourblind friendliness.
}
