% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brewerPalLength}
\alias{brewerPalLength}
\title{RColorBrewer Palette Length}
\usage{
brewerPalLength(name)
}
\arguments{
\item{name}{a character string}
}
\value{
a numeric value
}
\description{
Get the maximum length of an RColorBrewer Palette
}
