% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{brewerPalNames}
\alias{brewerPalNames}
\title{Get RColorBrewer Palette Names}
\usage{
brewerPalNames()
}
\value{
a character vector containing all RColorBrewer palette names
}
\description{
Get RColorBrewer Palette Names
}
