% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourPal.R
\name{colourPal}
\alias{colourPal}
\title{Flexible RColorBrewer Palettes}
\usage{
colourPal(name = NULL, len = NULL, type = "seq", reverse = F,
  shuffle = F, alpha = NA)
}
\arguments{
\item{name}{a character string. Default: NULL}

\item{len}{desired length of output palette. Note that if the length is larger than the maximum number of colours in the palette, the maximum number of colours will be returned. Default: NULL}

\item{type}{a character string indicating the type of palette desired. One of 'seq', 'sequential', 'div', 'diverging', 'qual', or 'qualitative'. Default: 'seq'}

\item{reverse}{a boolean value indicating whether the palette should be returned in reverse order. Note that for diverging palettes, this will effectively restore the original palette order since the function by default reverses them (see also Description). Default: F}

\item{shuffle}{a boolean value indicating whether the palette should be shuffled prior to return. Default: F}

\item{alpha}{modify colour transparency. 'NA' or numeric value between 0 and 1. Default: NA}
}
\value{
a character vector of colours.
}
\description{
Retrieve an RColorBrewer Palette of desired type. The default palettes are 'YlOrBr' for sequential palettes, 'RdBu' for qualitative palettes and 'Dark2' for qualitative palettes. Options to reverse or shuffle the palette prior to returning. Note that in the default RColorBrewer configuration, the warmer colours in the divergent palettes are at the start rather than the end of the scales. In this function, all divergent palettes are by default reversed, such that the RdBu palette for example will go from Blue to Red rather than the original Red to Blue.
}
\seealso{
\code{\link[RColorBrewer]{RColorBrewer}}
}
