% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{isBrewerDiv}
\alias{isBrewerDiv}
\title{Diverging RColorBrewer Palettes}
\usage{
isBrewerDiv(name)
}
\arguments{
\item{name}{a character string}
}
\value{
a boolean value
}
\description{
Check if an RColorBrewer palette is diverging.
}
