% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{isBrewerPal}
\alias{isBrewerPal}
\title{Test for Valid RColorBrewer Palette Name}
\usage{
isBrewerPal(name)
}
\arguments{
\item{name}{a character string}
}
\value{
a boolean value
}
\description{
Test if a character string is a valid RColorBrewer palette name.
}
